reorganize Rails specifics
