+++ lib/bali.rb
@@ -10,7 +10,7 @@ end
 require "zeitwerk"
 loader = Zeitwerk::Loader.for_gem
 loader.ignore("#{__dir__}/generators")
-loader.ignore("#{__dir__}/bali/activerecord.rb")
+loader.ignore("#{__dir__}/bali/rails")
 loader.setup
 
 module Bali
@@ -32,7 +32,7 @@ module Bali
 
   if defined? Rails
     require "bali/railtie"
-    require "bali/activerecord"
+    require "bali/rails/active_record"
   end
 end
