+++ lib/bali/rails/active_record.rb
